Add revolving report data. This data only appears if its value is greater than zero. While I am at it,
Added a [*] to these so they can be picked out quickly when they turn on,
lets just hope they don't as they are vital warnings that a loss of data
has occurred.

Style updates to the code to make it more readable in a 80 column terminal.
This is mainly some printf wraps.

Tabulated the description column.