Switch pkg_commit(1) over to Mercurial by default.
By design subversion uses .svn directories in each directory that is under
its control where as hg or mercurial uses one location .hg in the root of
the repository. Using Subversion for a package database is not ideal for
keeping track of changes.

Support for git(1) will be added in the near future.