Add acltool(1).
This eases support for setting ACLs and storing them for use later
on when they might need to be restored due to upgrades etc...

All files in the /etc/acls directory are generated with getfacl(1)
and redirecting output to a file for editing after the new acl of the
file has already been set using setfacl(1).

This script also supports proccessing of the comments area for keyword
"chflags" with one argument of a comma seperated "no-spaces" list of
flags.

Example:
# chflags: schg,nodump,sappnd

Since the setting of ACLs depends on schg and or uchg flags not being
present these are removed prior to processing the .acl files located
in the /etc/acls/ directory and reset using the above example. Be sure
to properly edit your ACLs to add these flags if they are important for
your secure operations.

Exmaple .acl file:
#######################################################################
# file: /etc/acls
# owner: root
# group: wheel
# chflags: sunlnk
user::rwx
group::---
mask::---
other::---
#######################################################################